fix: merge sync-preflight contract tests for pkg-file-count limit
