+++ docs/internal/processes/api-relevant-files.md
@@ -104,7 +104,7 @@ Use this map when changing the public REST contract.
 - `pkg/api/contracttests/` contains OpenAPI authoring, bundled-contract, and generated-type smoke tests for public API changes. Residual active-surface import gate verification uses scoped `rg` commands documented in `docs/internal/development/plans/dynamic-workflows/dynamic-workflows-plan.md` plus behavioral preview/alias tests in contract and server packages.
 - Internal provider/session response-stream changes stay out of the public contract by proving absence at the artifact boundary in `pkg/api/contracttests/`: when adding provider-specific normalization or internal `SessionResponseStream` semantics, add or update contract tests against `api/openapi.yaml`, `pkg/api/generated/server.gen.go`, `pkg/generatedclient/client.gen.go`, `ui/src/api/generated/openapi.ts`, and canonical public event fixtures so internal response-stream terms and provider-native event names do not leak into published schemas or `FactoryEvent` payload artifacts.
 - Durable factory-session contract fixtures live in `pkg/api/testdata/durable-session-contract-fixtures.json` with round-trip and OpenAPI validation coverage in `generated_contract_durable_session_test.go`; keep `WorkContent` payloads as ordered part arrays and use API-relative artifact retrieval refs instead of host filesystem paths.
-- Session sync-preflight recovery contract fixtures live in `pkg/api/testdata/sync-preflight-recovery-contract-fixtures.json` with OpenAPI validation, generated Go round-trip, and `sessionpersistence.ClassifyIdentityMismatch` coverage in `pkg/api/contracttests/generated_contract_sync_preflight_recovery_test.go`; prove typed HTTP `200` recovery outcomes for `ok`, `cursor_stale`, `session_not_found`, and `logical_session_remap` in `pkg/service/runtime_session_runtime_test.go`, `pkg/api/servertests/server_factory_session_orchestrator_test.go`, and `ui/src/api/factory-sessions/api.test.ts` instead of relying on transport failures for stale or missing sessions.
+- Session sync-preflight recovery contract fixtures live in `pkg/api/testdata/sync-preflight-recovery-contract-fixtures.json` with OpenAPI validation, generated Go round-trip, and `sessionpersistence.ClassifyIdentityMismatch` coverage in `pkg/api/contracttests/openapi_contract_events_test.go`; prove typed HTTP `200` recovery outcomes for `ok`, `cursor_stale`, `session_not_found`, and `logical_session_remap` in `pkg/service/runtime_session_runtime_test.go`, `pkg/api/servertests/server_factory_session_orchestrator_test.go`, and `ui/src/api/factory-sessions/api.test.ts` instead of relying on transport failures for stale or missing sessions.
 - Adding a new OpenAPI operation updates `factoryapi.ServerInterface`; add or update the matching `pkg/api` handler before running package tests.
 - Managed-runtime discovery and inspect projection live in `pkg/localmodels/catalog.go` and `pkg/localmodels/managed_runtime.go`; wire runtime cache and source resolver inputs through `localmodels.CatalogOptions` from `pkg/models/service` and `pkg/service/model_catalog.go`. Process-wide local model ownership belongs in `pkg/modelhost` (`Host`, `CatalogHost`, `ManagedRuntimeFromSnapshot`); construct the host at service build time through `newRuntimeLocalModelDependencies` in `pkg/service/factory_build.go` and expose it from `FactoryCore.ModelHost()`. Model list, detail, pull, and direct invocation ownership belongs in `pkg/models/service` (`Service` implements `apisurface.ModelAPI` via `NewFromHost`); wire production collaborators through `pkg/service/model_catalog.go` (`modelServiceHost`, `wireModelServiceCollaborator`, `ProvideModelServiceCollaborator`, `AttachModelServiceCollaborator`) and keep the same file as a compatibility facade that delegates to the injected `apisurface.ModelAPI`. `pkg/localmodels` continues to own catalog structure projection and invocation-readiness classification behind `pkg/modelhost`. Supervised `LLAMACPP` load/readiness/crash handling lives in `pkg/modelhost/supervisor.go`, `pkg/modelhost/llamacpp_supervisor.go`, `pkg/modelhost/process_launcher.go`, and `pkg/modelhost/health.go`; worker configs must include `--health-endpoint` for supervised server startup. Shared lease capacity, idle unload, and resource-pressure eviction live in `pkg/modelhost/lease_policy.go` with `Options.IdleUnloadAfter` and `Options.MaxLoadedRuntimes`. Model host operator diagnostics (logs and `model_host.*` counters) live in `pkg/modelhost/diagnostics.go` with zap/metrics wiring in `pkg/service/modelhost_diagnostics.go`; see `docs/architecture/model-host.md`.
 - Managed-runtime pull projection and outcome normalization live in `pkg/localmodels/pull_lifecycle.go` and `pkg/localmodels/catalog.go` (`PullModelWithOptions`); API transport mapping for `managedRuntimePull` belongs in `pkg/apisurface/managed_runtime.go`; pull logging and `managed_runtime.pull.*` metrics belong in `pkg/models/service/pull.go` with pull-metrics adaptation at `pkg/service/model_catalog.go`.
