+++ pkg/api/contracttests/openapi_contract_events_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/sessionpersistence"
 )
 
 func TestOpenAPIContract_DefinesUnifiedFactoryEventLog(t *testing.T) {
@@ -668,4 +669,305 @@ func assertJSONKeysPresent(t *testing.T, object map[string]any, name string, key
 			t.Fatalf("%s.%s is missing", name, key)
 		}
 	}
+}
+
+type syncPreflightRecoveryFixtureCatalog struct {
+	Scenarios                []syncPreflightRecoveryScenario      `json:"scenarios"`
+	IdentityScopeComparisons []syncPreflightIdentityScopeScenario `json:"identityScopeComparisons"`
+}
+
+type syncPreflightRecoveryScenario struct {
+	ID       string                    `json:"id"`
+	Tags     syncPreflightRecoveryTags `json:"tags"`
+	Response map[string]any            `json:"response"`
+}
+
+type syncPreflightRecoveryTags struct {
+	ReasonCode         string `json:"reasonCode"`
+	CheckpointReusable string `json:"checkpointReusable,omitempty"`
+	CursorValid        string `json:"cursorValid,omitempty"`
+}
+
+type syncPreflightIdentityScopeScenario struct {
+	ID                 string         `json:"id"`
+	Previous           map[string]any `json:"previous"`
+	Current            map[string]any `json:"current"`
+	WantClassification string         `json:"wantClassification"`
+}
+
+func TestOpenAPIContract_SyncPreflightRecoveryFixturesValidateAndRoundTrip(t *testing.T) {
+	doc := loadValidatedOpenAPIContract(t)
+	catalog := loadSyncPreflightRecoveryFixtureCatalog(t)
+
+	seenReasonCodes := map[string]struct{}{}
+	for _, scenario := range catalog.Scenarios {
+		t.Run(scenario.ID, func(t *testing.T) {
+			assertSyncPreflightRecoveryScenarioFixture(t, doc, scenario)
+			seenReasonCodes[scenario.Tags.ReasonCode] = struct{}{}
+		})
+	}
+
+	for _, reasonCode := range []string{"ok", "cursor_stale", "session_not_found", "logical_session_remap"} {
+		if _, ok := seenReasonCodes[reasonCode]; !ok {
+			t.Fatalf("sync preflight recovery fixture coverage for %s = missing, want scenario", reasonCode)
+		}
+	}
+}
+
+func TestOpenAPIContract_SyncPreflightIdentityScopeComparisonsDistinguishBackendAndStreamChanges(t *testing.T) {
+	catalog := loadSyncPreflightRecoveryFixtureCatalog(t)
+
+	for _, scenario := range catalog.IdentityScopeComparisons {
+		t.Run(scenario.ID, func(t *testing.T) {
+			previous := identityScopeFromFixtureMap(scenario.Previous)
+			current := identityScopeFromFixtureMap(scenario.Current)
+
+			reason, ok := sessionpersistence.ClassifyIdentityMismatch(previous, current)
+			if !ok {
+				t.Fatal("ClassifyIdentityMismatch = false, want mismatch")
+			}
+			if string(reason) != scenario.WantClassification {
+				t.Fatalf("classification = %q, want %q", reason, scenario.WantClassification)
+			}
+
+			if previous.BackendScopeID == current.BackendScopeID &&
+				scenario.WantClassification == string(sessionpersistence.ReasonBackendScopeChanged) {
+				t.Fatal("backend scope classification requires backendScopeId change")
+			}
+			if previous.StreamGenerationID == current.StreamGenerationID &&
+				scenario.WantClassification == string(sessionpersistence.ReasonStreamGenerationChanged) {
+				t.Fatal("stream generation classification requires streamGenerationId change")
+			}
+			if previous.BackendScopeID != current.BackendScopeID &&
+				scenario.WantClassification == string(sessionpersistence.ReasonStreamGenerationChanged) &&
+				previous.FactorySessionID == current.FactorySessionID {
+				if previous.BackendScopeID == current.BackendScopeID {
+					t.Fatal("stream-only classification should not change backendScopeId")
+				}
+			}
+		})
+	}
+}
+
+// Gate evidence for session-persistence-hardening-and-observability-005: proves the
+// verification gates still protect observable recovery behavior, not only compile-time
+// contract shape.
+func TestSessionPersistenceHardeningGateEvidence_RecoveryOutcomesControlCheckpointReuse(t *testing.T) {
+	doc := loadValidatedOpenAPIContract(t)
+	catalog := loadSyncPreflightRecoveryFixtureCatalog(t)
+
+	var reusableScenario *syncPreflightRecoveryScenario
+	var staleScenario *syncPreflightRecoveryScenario
+	for index := range catalog.Scenarios {
+		scenario := catalog.Scenarios[index]
+		switch scenario.Tags.ReasonCode {
+		case "ok":
+			reusableScenario = &scenario
+		case "cursor_stale":
+			staleScenario = &scenario
+		}
+	}
+	if reusableScenario == nil || staleScenario == nil {
+		t.Fatal("sync preflight recovery fixtures missing ok or cursor_stale scenario")
+	}
+
+	assertSyncPreflightRecoveryScenarioFixture(t, doc, *reusableScenario)
+	assertSyncPreflightRecoveryScenarioFixture(t, doc, *staleScenario)
+
+	var reusableResponse generated.FactorySessionSyncPreflightResponse
+	assertGeneratedFixtureRoundTrip(t, reusableScenario.Response, "FactorySessionSyncPreflightResponse", func(raw []byte) {
+		decodeRoundTripJSON(t, raw, &reusableResponse, reusableScenario.ID+" reusable response")
+	})
+	var staleResponse generated.FactorySessionSyncPreflightResponse
+	assertGeneratedFixtureRoundTrip(t, staleScenario.Response, "FactorySessionSyncPreflightResponse", func(raw []byte) {
+		decodeRoundTripJSON(t, raw, &staleResponse, staleScenario.ID+" stale response")
+	})
+
+	if !reusableResponse.CheckpointReusable {
+		t.Fatal("reusable recovery outcome checkpointReusable = false, want true")
+	}
+	if staleResponse.CheckpointReusable {
+		t.Fatal("stale cursor recovery outcome checkpointReusable = true, want false")
+	}
+
+	staleDiagnostic, ok := sessionpersistence.InvalidationFromSyncPreflight(staleResponse)
+	if !ok {
+		t.Fatal("stale cursor recovery outcome missing invalidation diagnostic")
+	}
+	if staleDiagnostic.Reason != sessionpersistence.ReasonCursorStale {
+		t.Fatalf("stale cursor diagnostic reason = %q, want %q", staleDiagnostic.Reason, sessionpersistence.ReasonCursorStale)
+	}
+	if staleDiagnostic.RecoveryAction != sessionpersistence.RecoveryReplayWithoutCursor {
+		t.Fatalf(
+			"stale cursor diagnostic recovery = %q, want %q",
+			staleDiagnostic.RecoveryAction,
+			sessionpersistence.RecoveryReplayWithoutCursor,
+		)
+	}
+}
+
+func assertSyncPreflightRecoveryScenarioFixture(
+	t *testing.T,
+	doc *openapi3.T,
+	scenario syncPreflightRecoveryScenario,
+) {
+	t.Helper()
+
+	assertOpenAPIFixtureValidates(t, doc, "FactorySessionSyncPreflightResponse", scenario.Response)
+	assertGeneratedFixtureRoundTrip(t, scenario.Response, "FactorySessionSyncPreflightResponse", func(raw []byte) {
+		var value generated.FactorySessionSyncPreflightResponse
+		decodeRoundTripJSON(t, raw, &value, scenario.ID+" sync preflight response")
+		assertSyncPreflightRecoveryOutcome(t, scenario, value)
+	})
+}
+
+func assertSyncPreflightRecoveryOutcome(
+	t *testing.T,
+	scenario syncPreflightRecoveryScenario,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	if string(response.ReasonCode) != scenario.Tags.ReasonCode {
+		t.Fatalf("%s reasonCode = %q, want %q", scenario.ID, response.ReasonCode, scenario.Tags.ReasonCode)
+	}
+
+	switch response.ReasonCode {
+	case generated.Ok:
+		assertSyncPreflightOkOutcome(t, scenario.ID, response)
+	case generated.CursorStale:
+		assertSyncPreflightCursorStaleOutcome(t, scenario.ID, response)
+	case generated.SessionNotFound:
+		assertSyncPreflightSessionNotFoundOutcome(t, scenario.ID, response)
+	case generated.LogicalSessionRemap:
+		assertSyncPreflightLogicalSessionRemapOutcome(t, scenario.ID, response)
+	default:
+		t.Fatalf("%s reasonCode = %q, want supported recovery outcome", scenario.ID, response.ReasonCode)
+	}
+
+	assertSyncPreflightInvalidationDiagnostic(t, scenario.ID, response)
+}
+
+func assertSyncPreflightOkOutcome(
+	t *testing.T,
+	scenarioID string,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	if !response.CheckpointReusable {
+		t.Fatalf("%s checkpointReusable = false, want true", scenarioID)
+	}
+	if !response.ReconnectCursor.ValidForStreamGeneration {
+		t.Fatalf("%s reconnect cursor validForStreamGeneration = false, want true", scenarioID)
+	}
+	if response.BackendScopeId == nil || response.FactorySessionId == nil || response.StreamGenerationId == nil {
+		t.Fatalf("%s identity fields = %#v, want full identity set", scenarioID, response)
+	}
+}
+
+func assertSyncPreflightCursorStaleOutcome(
+	t *testing.T,
+	scenarioID string,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	if response.CheckpointReusable {
+		t.Fatalf("%s checkpointReusable = true, want false", scenarioID)
+	}
+	if response.ReconnectCursor.ValidForStreamGeneration {
+		t.Fatalf("%s reconnect cursor validForStreamGeneration = true, want false", scenarioID)
+	}
+	if response.BackendScopeId == nil || response.FactorySessionId == nil || response.StreamGenerationId == nil {
+		t.Fatalf("%s identity fields = %#v, want full identity set for stale cursor", scenarioID, response)
+	}
+}
+
+func assertSyncPreflightSessionNotFoundOutcome(
+	t *testing.T,
+	scenarioID string,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	if response.CheckpointReusable {
+		t.Fatalf("%s checkpointReusable = true, want false", scenarioID)
+	}
+	if response.BackendScopeId != nil || response.FactorySessionId != nil || response.StreamGenerationId != nil {
+		t.Fatalf("%s identity fields = %#v, want nil for missing session", scenarioID, response)
+	}
+}
+
+func assertSyncPreflightLogicalSessionRemapOutcome(
+	t *testing.T,
+	scenarioID string,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	if response.CheckpointReusable {
+		t.Fatalf("%s checkpointReusable = true, want false", scenarioID)
+	}
+	if response.BackendScopeId == nil || response.FactorySessionId == nil || response.StreamGenerationId == nil {
+		t.Fatalf("%s identity fields = %#v, want full identity set for remap", scenarioID, response)
+	}
+}
+
+func assertSyncPreflightInvalidationDiagnostic(
+	t *testing.T,
+	scenarioID string,
+	response generated.FactorySessionSyncPreflightResponse,
+) {
+	t.Helper()
+
+	diagnostic, ok := sessionpersistence.InvalidationFromSyncPreflight(response)
+	switch response.ReasonCode {
+	case generated.Ok:
+		if ok {
+			t.Fatalf("%s invalidation diagnostic = %#v, want none for ok", scenarioID, diagnostic)
+		}
+	default:
+		if !ok {
+			t.Fatalf("%s invalidation diagnostic missing for %q", scenarioID, response.ReasonCode)
+		}
+	}
+}
+
+func identityScopeFromFixtureMap(payload map[string]any) sessionpersistence.IdentityScope {
+	return sessionpersistence.IdentityScope{
+		BackendScopeID:      stringFixtureValue(payload, "backendScopeId"),
+		LogicalSessionKeyID: stringFixtureValue(payload, "logicalSessionKeyId"),
+		FactorySessionID:    stringFixtureValue(payload, "factorySessionId"),
+		StreamGenerationID:  stringFixtureValue(payload, "streamGenerationId"),
+	}
+}
+
+func stringFixtureValue(payload map[string]any, key string) string {
+	value, ok := payload[key].(string)
+	if !ok {
+		return ""
+	}
+	return value
+}
+
+func loadSyncPreflightRecoveryFixtureCatalog(t *testing.T) syncPreflightRecoveryFixtureCatalog {
+	t.Helper()
+
+	fixtureBytes, err := os.ReadFile("../testdata/sync-preflight-recovery-contract-fixtures.json")
+	if err != nil {
+		t.Fatalf("read sync preflight recovery contract fixtures: %v", err)
+	}
+
+	var catalog syncPreflightRecoveryFixtureCatalog
+	if err := json.Unmarshal(fixtureBytes, &catalog); err != nil {
+		t.Fatalf("decode sync preflight recovery contract fixtures: %v", err)
+	}
+	if len(catalog.Scenarios) == 0 {
+		t.Fatal("sync preflight recovery contract fixtures contain no scenarios")
+	}
+	if len(catalog.IdentityScopeComparisons) == 0 {
+		t.Fatal("sync preflight recovery contract fixtures contain no identity scope comparisons")
+	}
+	return catalog
 }
